requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
anyjson==0.3.3
asgiref==3.4.1
astroid==2.12.9
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==4.4.2
Deprecated==1.2.13
dill==0.3.5.1
dj-database-url==0.5.0
Django==3.2.9
django-autocomplete-light==3.9.4
django-choices==1.7.2
django-ckeditor==6.4.2
django-countries==7.3.2
django-country2==0.0.8
django-crum==0.7.9
django-currentuser==0.5.3
django-debug-toolbar==3.2.2
django-dotenv==1.4.2
django-environ==0.8.1
django-filter==22.1
django-heroku==0.3.1
django-jazzmin==2.4.9
django-js-asset==2.0.0
django-model-utils==4.2.0
django-modeladmin-reorder==0.3.1
django-mptt==0.13.4
django-nested-admin==3.4.0
django-nested-inline==0.4.4
django-rest-passwordreset==1.2.1
django-seed==0.3.1
django-sql-utils==0.6.1
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==13.15.1
frozenlist==1.3.0
future==0.18.2
geocoder==1.38.1
geoip2==4.6.0
gprof2dot==2021.2.21
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
imageio==2.14.0
imageio-ffmpeg==0.4.5
inflection==0.5.1
install==1.3.5
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
maxminddb==2.2.0
mccabe==0.7.0
moviepy==1.0.3
multidict==6.0.2
numpy==1.22.1
oauth2client==4.1.3
openapi-codec==1.3.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.0.0
platformdirs==2.5.2
proglog==0.1.9
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.5.0
pylint==2.15.2
pylint-plugin-utils==0.7
pyparsing==3.0.6
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2021.3
pyzmq==23.1.0
ratelim==0.1.6
redis==4.3.4
requests==2.26.0
rsa==4.9
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
statprof==0.1.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
toposort==1.7
tqdm==4.62.3
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
yarl==1.7.2
